Updating build targets and debug app
Assuming the production app will not need a proxy, this change makes one single 'war' target for deploying the app.  To simplify the debug/production setup, templates are removed in favor of static index.html files.  This includes some duplication of markup, but this is preferable to having the dynamic back-end where it is not needed.